+++ CHANGELOG.md
@@ -97,4 +97,4 @@ Start to build my personal js-core lib.
 -	增加跨线程通道tunnel
 
 ### 2018.11.15
--	增加Channel的合并功能
+-	增加Channel和Tunnel的合并功能